custom_extensions/requirements.txt
openai>=1.30.0
azure-search-documents>=11.4.0
azure-identity>=1.17.0
streamlit>=1.35.0
pandas>=2.0.0
